Install events library from npm
Bundlers like vite does not automatically polyfill node modules,
functionPlot uses the events library from npm and relies on the
bundler to bundle it when needed.

The solution is to install events from npm, it has the same API
and it tells the bundler that it should install it when functionPlot
is used.